Automatically backed up by Learn

+++ index.js
@@ -28,6 +28,6 @@ function makeInt(n) {
 }
 
 function preserveDecimal(n) {
-  parseInt(n)
+  parseFloat(n)
   return n
 }
